requirements.txt
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
fonttools==4.53.0
imageio-ffmpeg==0.5.1
kiwisolver==1.4.5
matplotlib==3.9.0
multiprocess==0.70.16
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==70.0.0
six==1.16.0
tzdata==2024.1
xarray==2024.6.0
